feat: Create NewsletterDetail component with image and reference link handling
